maps_reviews/requirements.txt
googlemaps>=4.10.0
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26.0
tqdm>=4.66.0
matplotlib>=3.9.0
scipy>=1.11.0
statsmodels>=0.14.2
snownlp>=0.12.3
pyarrow>=17.0.0
